Gabriel is correct. His first set of changes were at least not wholly causing the SEGVs.  We have an *existing* problem using the NonUniformDegreeTreeChildIterator with Unlinks().
I have introduced a 'fix', but it really needs to be replaced because
it significantly slows down normalization.

Consider this snippet of code:
<L b=... >  // level 3
  ...
  <L b="1484" e="1485">  // level 4
     <S b="1484" />
     <S b="1484" />
     <S b="1485" />
     <S b="1485" />
  </L>
  <S b="1485">
  ...
</L>

CDS_Main loops on the children of the current scope.  It saves the
current value of the iterator to 'child' and then increments it (an
attempt to prevet this problem!). After that it recursively calls CDS
on 'child'.

When 'child' points to the 4th level loop in the picture above, the
iterator has been incremented and points to the next child which
happens to be a statement for line 1485.

When CDS processes the inner loop and reaches a statement for line
1485, it realizes a duplicate was found in the level 3 loop -- the
statement to which level 3 loop iterator points to.  Because the
current statement is on a deeper level (4 as opposed to 3 in the map),
the statement on level 3 is unlinked and deleted.  After the level 4
loop is completely processed, it returns back to the level 3 loop
where it points to a block of memory that was freed already. The
program dies on a dynamic_cast of this pointer.

If you ask why the level 3 statement to 1485 has already been
processed when CDS is looking within the level 4 loop, the answer is
that a loop fusion has already taken place and CDS was restarted from
the least common ancestor.