Phase 8 of ScopeInfo.* and PgmScopeTree.* merge: Everything is merged except PerfMetric stuff.
One question: Within ScopeInfo.cpp, the fuction CodeInfo::SetLineRange
had a call to "Relocate()" that was commented out:

      if (changed && (mom != NULL)) {
  #if 0 // this doesn't achieve the result we want
        Relocate(); 
  #endif
        mom->SetLineRange(begLine, endLine); 

I could neither think of a reason why uncommenting it would be a
problem nor detect any malfunction after a quick test.  Thus, in order
to get the penultimate stage of the merge, I will uncomment it.