When merging profiles that have conceptually, but not technically, equal load maps, hpcprof generates databases that have unprocessed CALLSITEs.
A specific example is s3d, whose 'build system' kindly prepares run directories for different inputs, complete with a *copy* of the executable.  Thus, runs for two different inputs technically use distinct binaries with different names (paths).

When trying to process profiles from two different runs (for a scaling study), the profiles must be merged.  As the epochs are merged, Epoch::merge determines that the current epoch has no corresponding module for the binary (because of the different paths).  It then supplements the current load map with the missing version of the module.  Conceptually, everything should be fine because this procedure preserves the link each node in the CCT has with each version of the module.

The difficulty arises in how the Prof::Epoch is represented.  An Epoch contains indices by id, name and VMA for accessing its modules (Prof::Epoch::LM).  The name index is a multiset because the same module can be mapped to different areas of the address space.  However, the VMA index is a set because it is necessary to have a strict partition of the address space so that call sites can be associated with their respective load modules.  In the case above, the s3d executable has a different name (path) but the same load address.  So when the additional module is merged into the Epoch (Epoch::lm_insert), a reference is inserted in the name index, but cannot be inserted in the VMA index because a direct conflict exists.

The final piece of the problem is in the hpcprof driver which iterates over the modules in the Prof::Epoch for the purpose of creating procedure frames.  To iterate in a nice and orderly fashion, I used the VMA index -- the index missing the additional module.  The results was that any call site linked to the additional module was not processed, since the module was never processed.

To enable hpcprof to handle the situation described I have:
  - changed the hpcprof driver to use the id index instead of the VMA index
  - inserted a warning to detect this situation
  - beefed printing/debugging code (which always occurs in a demand-driven fashion)

As a final comment, it is not clear that we want hpcprof to support such usage, at least at the binary correlation level.  In addition to the difficulty of aligning load maps, we would need to ensure that any structure information is aligned as well.  (With the current change, STRUCTURE information is only used for the module that exactly matches.)  Since we need source-level comparisons anyway, it seems more desirable for this support to address studies across different binaries.