- When computing branch targets, use the new function normalizeTarget() to compute the closest *actual* target VMA.  Technically, this should never be necessary, because decoding should be correct.  However, x86 has a terrible penchant for creating new instructions and we are not quite keeping up with binutils.  This little bit of trickery should be enough to hide the fact that binutils is botching a few decodes.
The particular motivation for this fix was Fowler's examination of Joo's QCD codes.  binutils failed to decode a nopl instruction (apparently related to virtualization) which created a CFG with unreachable basic blocks where an important loop should have been.  This fix has made a big difference in the structure information.

- OAInterface::normalizeTarget(): create.

- A bit of cleanup.