Sometimes we are hammered in the middle of a Cilk operation where the cactus stack pointer becomes NULL, even though logical context 'should' exist (e.g., Cilk_sync before a return).  The result is that we cannot obtain the logical context and therefore cannot locate the unwind correctly within the CCT.
To avoid such bad unwinds, we use the simple heuristic of
requiring a cactus stack pointer if the worker has become a
thief. It is not a perfect solution since a worker may have
stolen the main routine; however, this is an exceptional case.

FIXME: the best solution is to find the smallest window within
the scheduler code and set a flag.  But I found that this is
easier said than done.