Part one of the "safe sampling" code.  This reworks the sync vs. async and suspend sampling rules.
1. Ignore monitor callbacks before csprof is initialized (at main).
Monitor promises that init and fini process and thread occur in
sequence, but dlopen(), fork(), etc, can occur at any time, and this
causes trouble if they happen before main.

We still have trouble in the child process if fork() happens before
main, but we no longer segfault.

2. Block all async entry (via profiling signal) while already in
csprof synchronously (via monitor callback).  This is dangerous in
general, and can cause deadlock via the MSG lock.  This change
supersedes the handling_synchronous_sample thread data.

Synchronous monitor callbacks are infrequent and short, so this blocks
only a few samples.  The control variable ASYNC_RISKY will disable
this block and allow sampling inside monitor callbacks, but that runs
the risk of deadlock.

3. Report the number of blocked samples.  Note: this patch doesn't
increase the number of blocked samples (don't panic), but we now
report them.

4. If argv is NULL in monitor_init_process(), which happens if fork()
is called before main, then try to find the process name via /proc.
This is mostly a work-around for monitor so we don't segfault.

Also, block sampling while inside dlopen().  This is a temporary,
thread-specific solution for now.

Add the control variables ASYNC_RISKY (as above) and DLOPEN_RISKY to
disable the safety features, if desired.