For x86 unwind: Check bp for sanity before beginning a bp unwind    (sanity means sp <= bp < monitor_stack_bottom)  This cuts down on segv.
Also, change pointer arithmetic to use 'void *' rather than 'unsigned long'

Preliminary validation for trolling is checked in, but it is not active.
The old method of checking for function bounds is still the active troll validation method.

Make suspicious interval logging use TMSG instead of EMSG. It is turned on by default, but
now it can be turned off.