Part two of the "safe sampling" code.  This part addresses dlopen and friends.
Install a readers-writers lock around dlopen(), dlclose() and
dl_iterate_phdr(), that is, one writer and no readers, or else no
writers and any readers.  dlopen and dlclose modify the load map, so
they're writers, and dl_iterate_phdr (via sampling or pthread_create)
reads the load map.  This lock is process-wide.

The control variable DLOPEN_RISKY bypasses the read lock, that is, it
allows sampling during dlopen.  This is risky and could potentially
deadlock or segfault, but usually the window of vulnerability is very
narrow.  Taking the risk allows sampling during init constructors.
(And if we could just separate dlopen into mmap and the init
constructor, then we could get the best of both worlds, safety and
profiling init constructors.)

This was supposed to be the last of the "safe sampling" code, but
testing has revealed further issues and race conditions.

(1) There is a lock-order-reversal deadlock between the fnbounds lock
and the system dl lock.  This is currently avoided by having the
dlopen lock extend a bit farther than it should.  (It should drop to a
read lock while in fnbounds map/unmap dsos.)

(2) There is a potential end-of-thread deadlock that I haven't tracked
down yet.

Move the number-of-samples variables into sample_event.c and make
static, along with the end-of-process summary, and export increment
and count functions as needed.