Add the infrastructure for monitor extensions, that is, the ability to override library functions beyond what libmonitor provides.  Use this to override pthread mutex locks, spin locks and cond wait.
Currently, lush is the only client for the monitor extensions.  Add
configure options for turning on the monitor and lush extensions with
default no.

    --enable-monitor-exts
    --enable-lush-pthreads

Note: the monitor extensions are somewhat broken in the static case
and will segfault if some debugging options are turned on.  This can
be fixed by reworking some things either on the libmonitor or
hpctoolkit side.

Add the locks.c test program that makes some simple calls to mutex
locks, spin locks and cond wait.