hpcrun and hpcrun-fmt: A "bit" of reorganization to support extending a loadmap entry to have an id.  (Other famously simple tasks: Let's just all get along.)
Background: It turns out that there were two nearly idential
declarations for a loadmap and loadmap entry within hpcrun-fmt: the
reader read one version the writer wrote the other.  While I
understand that this was done to avoid copying a structure in hpcrun,
the reader/writer should probably use the same data structure.