Add initial and empty overrides for longjmp() and siglongjmp(). These are used by the trampoline.
Remove the --enable-monitor-exts option.  The overrides of the
mutex, spin and cond wait functions are now controlled solely by
the --enable-lush-pthreads option.

Remove building libcsprof_wrap.a (for now) which was basically empty.
Library functions not in libc should be overridden in separate .o
files (in the static case) to avoid always pulling in that library,
so we may readd this at some point.