In the hpclink script: the extra undef args should be "-Wl,-u,${name}", not "-Wl,-u,__wrap_${name}".  This was an outright bug and broke the -u workaround.
Add a workaround for gfortran: if the first link fails with undefined
reference to 'main', then force main to be undefined.  We already do
this with undefined __wrap_ symbols.

Change the example /path/to/prefix to the actual prefix, in case
someone wants to move and modify the script.

Thanks to: Andrew Porter at STFC Daresbury Laboratory (UK) for
reporting the problem.