In LM.cpp, fetch the synthetic symbol table and append it to the regular table in our local sorted copy.  On many platforms, the synthetic table is empty, but not on power7.  By default, 64-bit binaries on power7 classify function symbols as 'D' (initialized data) instead of 'T' (text), and this patch helps identify the symbol name instead of just '.text'.
Also, fix a couple 'new' allocations.  The binutils funtions return
the table sizes as bytes (the correct number for malloc), but new uses
number of array entries, not bytes.

Thanks to: Anton Blanchard of IBM for the idea of using the synthetic
table and for providing an earlier version of the patch.