hpcprof/hpcprof-mpi: normalize hpctrace files so they are consistent with the canonical CCT. Intermediate checkin.
Rearrange declarations of CCT::Tree and CCT::ANode so that
CCT::ANode::mergeMe can cleanly pass CCT::ANode::MergeEffects to outer
level functions.

LoadMapMgr::merge and LoadMap::merge really should return a pointer to
a vector rather than an actual vector.  (Since the return vector could
never be very large, this was never a big deal. But it still was a
mistake.)