Move the assembler .S files to .s files.  The .S files are run through the C preprocessor which causes problems with gcc 3.x.
The solution involves some automake hoops.  We have to redefine
CCASFLAGS for the hpcrun .o and .so targets so that the tramp.s files
will be properly renamed. (yuck)  See section 8.3.9.2, Objects created
with both libtool and without, in the automake docs.

Thanks to: Martin Burtscher at Univ Texas for reporting the problem.