Check the Fortran symbol using MPIF77 variable instead of F77. The reason is that it is possible that F77 and MPIF77 point to different compiler; thus the configure will incorrectly check the Fortran symbol.
Note: For MPI wrapper prototype, hpctoolkit is tightly coupled with MPI compiler. If the MPI Fortran compiler generates the fortran symbol with one underscore, then we wrap all MPI functions with one underscore. 
If the user uses different compiler (with two underscores for instance), then the wrapper doesn't work.

It the future we will try to cope this issue by checking the symbol during the initialization of hpcrun (for instance).