[Analysis::CallPath::]coalesceStmts(): Fix an overaggressive assertion while merging CCT::Stmts within a CCT::ProcFrm.
Because (a) trace records contain a function's representative IP
and (b) two traces that contain samples from the same function
should have their conflict resolved in Prof::CallPath::Profile::merge(),
we would expect that merge effects are impossible.  That is, we
expect that it is impossible that a CCT::ProcFrm has multiple
CCT::Stmts with distinct trace ids.

However, merge effects are possible *after* static structure is
added to the CCT.  The reason is that multiple object-level
procedures can map to one source-level procedure (e.g., multiple
template instantiations mapping to the same source template or
multiple stripped functions mapping to UnknownProcNm).