hpcprof-mpi:
Resolve a bug where trace ids are lost when hpcprof-mpi is run such
that an MPI process is assigned only one trace file.  We did not
notice this until now because, unlike the bug reporter, we typically
do not run hpcprof-mpi with the same number of MPI ranks as the
profiled application.

To fix this bug, I simply a statement that cleared a
CallPath::Profile's traceFileNameSet after *.hpcrun files were read.
I believe this statement was an outdated hack that I forgot to remove.
As far as I can see, the downstream code is now general enough that
this special case side-effect is no longer needed.