rename "compact" recursion support to "collapse" recursion improve support for collapsing recursion so that when the recursive chain ends and some other function is entered, the callsite where the other function is entered will always be entered into the tree as well.
thus, if there was a chain
main
  qsort - 1
    qsort -2 
      qsort - 3
        qsort - 4
          isort

the call chain would be compressed to

main
  qsort - 1
    qsort -2 
      qsort - 4
        isort

by keeping qsort - 4, we keep the call site for isort. in teh mpbs example, that means we can see where
isort and swap64 calls originate.

However, for fib, for
main
  fib - 1
    fib -2 
      fib - 3
        fib - 4
          fib line 6 

our approach would yield

main
  fib - 1
    fib -2 
      fib line 6 

we keep fib - 2 (and its sibling call) because that way we can see how much cost is incurred in each subtree of the
recursion.  there is no need to keep fib - 4 as we do for qsort above as it doesn't serve any purpose here.