Add a simple API to allow the application to start and stop sampling with function calls in the application source.
   void hpctoolkit_sampling_start(void);
   void hpctoolkit_sampling_stop(void);

Add a compatibility library, libhpctoolkit.{a,so}, with no-op
definitions of the start and stop functions so that an application can
link natively without having to #ifdef all of the calls.

Sample sources continue to run while stopped, we just don't unwind and
don't record samples.

By default, sampling starts on.  If the application wants it initially
off, then set HPCRUN_DELAY_SAMPLING in the environment, or use
--delay-sampling with hpcrun.

Although the API is fully functional, there are parts that require a
bit more manual effort than we would like.  For example, setting
LD_LIBRARY_PATH in the dynamic case.