In the fnbounds	code, where we create the command to send to the system server, add quotes around the file name to analyze.  This fixes the case where the file or directory name contains spaces.
Although we can	fix this case, it is not possible to handle every case
where a directory name contains spaces.  For example, it will never
work to install hpctoolkit into a directory name that contains spaces.
Hpcrun needs to add libraries from the install directory to LD_PRELOAD
and LD_PRELOAD uses space as a delimiter.  The practical and best
answer is don't use directory names with spaces on a Unix system.