In the PAPI sample source, treat failure in PAPI_library_init() as a soft error instead of a fatal error.  And delay printing an error message until after trying all other sample sources.
This is useful on ranger where PAPI is not available on the login
nodes, but the rest of hpctoolkit, including wallclock, still runs
on the login nodes.

There is no expectation that a single install of hpctoolkit should run
on both the login nodes and compute nodes (as it seems to on ranger).
But we can change this to a non-fatal error and try.