4 changes:
1) Biggest change:
Add a "only dump intervals" mode. To employ this feature, 
set the environment variable
    HPCRUN_ONLY_DUMP_INTERVALS=1,

and then either use hpcrun (dynamic), or invoke the program
directly (static).d

When in this mode, the output of this program is a dump of
all intervals in the fnbounds table of the executable.
*The executable itself does NOT run*.

NOTE: the output will be different, depending on whether
the user's program is statically or dynamically compiled.

2) The working directory used by the system server is now controlled by
the shell variable "TMPDIR". If "TMPDIR" is unset, the working directory
defaults to "/tmp".

3) Generalize the x86_dump_intervals and the ppc64_dump_intervals
   to a common interface "hpcrun_dump_intervals".
   This accomodates both the new "only dump intervals" mode as well as
   John Mellor-Crummey's special purpose single address
   interval dumping.

4) Rationalize temporary debugging code:
      4.1) Change "#if 0" to "#ifdef SPECIAL_DUMP_INTERVALS"
           This change allows inclusion of the special debug
           code when configured with
              CFLAGS="-DSPECIAL_DUMP_INTERVALS"
      4.2) Allow hex notation for addresses to have the intervals to be dumped.
           (Previous version insisted on decimal notation).
      4.3) Use general "hpcrun_dump_intervals" for
           portability on x86 and ppc64