Add the new fnbounds server.  We still create a pipe and fork a separate server process (so as not to run C++ code inside hpcrun). But instead of the server answering queries via system(), the server now execs hpcfnbounds and computes the answers itself.  This reduces the overhead of creating new processes from one per query to just one per process.
For now, we still run the old server by default.  To try the new
server, export NEW_SYSTEM_SERVER to the environment.

Some technical notes.

1. Write the array of addresses back over the pipe.  This eliminates
the need for temp files.

2. Automatically restart the server if it fails.

3. Catch segfaults to keep the server running if possible.

4. Catch SIGPIPE so that if the server crashes, it doesn't also kill
hpcrun.

5. There is a residual memory leak in the symtabAPI code.  It doesn't
seem to be too bad.  But if it becomes a problem, we could kill and
restart the server above some threshold (or else fix symtabAPI).

Some of the code is a bit clumsy in places while we continue to
support both servers.  This can be cleaned up when we eliminate the
old server.