Fix a cross-compile problem with the arch type for ld in hpcrun. Autoconf gets the wrong arch type for ld in a cross compile.  For example, on Intel MIC, autoconf uses 'ld -m elf_x86_64' instead of elf_k1om.
Fix this by attempting a partial link: 'ld -r'.  If that fails, then
use 'ld -V' to get a list of available arch types and retry until we
find one that works.

This may be a systemic problem in autoconf, or maybe it just doesn't
understand k1om for Intel MIC.  But the problem also exists in the
latest autoconf 2.69.