Change the inlined sequence of aliens in two ways.
1. Associate the alien cache with its physical parent instead of the
proposed scope.  This handles the case that a terminal stmt is inlined
through multiple paths better.

2. If a stmt and its proposed scope are both inlined, then start the
alien sequence at the child of the proposed scope.  The problem is
that proposed scope could be a loop in the middle of a long inlined
sequence.  Banal knows about the loop, but symtab doesn't and blows
past the loop all the way to the top-level, non-inlined code.