Modify the dlopen reader-writers lock to allow a writer thread to lock against itself in a nested dlopen.  That is, if a thread calls dlopen and the same thread calls dlopen from the init constructor, then we allow the thread to proceed.  This handles the case of calling dlopen from a library init constructor.  Previously, we could handle a library dependency correctly, but an explicit dlopen inside an init constructor would deadlock.
Normally, it would be dangerous to allow a thread to lock against
itself because that would expose an inconsistent state.  But for
dlopen, the dangerous part is already finished before starting the
init constructor.  And again, if we could just separate the init
constructor from the rest of dlopen, then we could handle this case
much better.