Fix two problems with the inline branch:
1. When locating a loop, stop the inline sequence at the scope for the
loop based on procnm.  The input inline sequence goes all the way to
the bottom code which is wrong for a loop.

2. When locating a node inside a scope, start the inline sequence at
the proposed scope.  Again, the input inline sequence goes all the way
to the outer procedure which is wrong for a middle scope (loop).  The
previous rev also did this, but this is a better way.