When working with LULESH 1.0.1 compiled -O3,  I found some cases where hpcstruct was reporting statements as being in a loop when no loop was present in the source code. A close examination of the assembly code showed that a combination of hot-path straightening combined with a tail call led to the appearance of a loop. Specifically, given an input program with the following structure
    S1
    if (cond1)
       call fn1
    S2
    ...
    call f2
    return

The compiler seems to have guessed that the call to fn1 was off the hot path and as a result
generated the following code 

     S1
     if cond1 goto l2
l1:  S2
     ...
     tail call to f2 instead of a return
l2:  inlined code for f1
     go to l1
 
Before this checkin, hpcstruct treated the branch to a location outside the routine as a SIMPLE 
statement. This led hpcstruct to erroneously conclude that there was a loop. With this revision,
we treat a branch to a code location outside the curent function as a tail call and correctly 
conclude that no loop is present.