hpcrun ompt branch: hack to fix the redundant appearance of the same nodes in the viewer
The problem occurs when profiling an openmp program with ompt.
Assume the master has the following bt: a -> b -> c -> fork()
the workers then (after eliding) will have the following bt: c -> d -> ...

Combining the two callpaths would result: a -> b -> c -> c -> d -> ...

If we eliminate the c in the worker we should avoid redundancy.

Note: we need to test with more complex applications.