Add a timeout for the time in minutes the server will wait to accept an incoming connection.  This is settable with the '--timeout num' option with a default of 15 min.
Since the server now accepts a series of connections, it makes sense
to timeout for security if not used.  This only applies to the time to
accept().  Once connected, the client can stay connected forever.

Also, display the IP address of the other end of the connection.
Note: if using an ssh tunnel, the address may appear as 127.0.0.1.