Add the command-line options --memsize and --chunksize to specify the amout of RAM for an in-memory cache of the mega-trace file and the cache page size.  The defaults are 50% of the available RAM for memsize and 6 Meg for chunksize.
It's difficult to automatically compute sensible defaults.  In the
serial case, we don't know if the machine is dedicated or multiuser.
And in the MPI case, we don't know how many ranks run on one node.
So, it's often best to give these on the command line.

Add the function Communication::rankLeader() that returns true if this
rank is the group leader, mostly for printing messages.