The first baby step to using parseAPI for control-flow graphs.  This much just cleans out two prereqs for using newer dyninst code.
1. Auto test for -std=c++11 and add it for the parseAPI code.  Due to
include files, this touches most of our C++ code, so we'll just commit
to it everywhere.

2. The latest dyninst has an incompatible change from 8.2-release over
how we get a function's demangled name in the struct inline code.  For
now, this requires a new dyninst from git.

Still a bit rough in places, and unfortunately, the dyninst code plus
boost give a spew of warnings with recent gcc.