Extend the ParseAPI traversal of functions, loops and blocks down to individual instructions and insert actual loop and stmt nodes into the scope tree.  We use the same location manager and reparenting algorithms as from before.  At this point, hpcstruct with ParseAPI produces an actual struct file and does so without using anything from Open Analysis.
However, there are still several parts that are incomplete or broken.
The most significant of these is there is a major regression in the
dyninst inline info from 8.2-release to git head.  Wisconsin is
working on the problem, but until it is fixed, the inline part of the
struct files will be badly broken.  This affects both the ParseAPI and
Open Analysis versions.