First draft enhancement to reduce procedure name redundancy in experiment.xml
By default, hpcprof will eliminate any procedures in procedure table if the name of the procedure is exactly the same and located in the same file.

For instance:
<procedure i=11 n="foo">
<procedure i=13 n="foo">

Procedure 13 will be removed if it is defined in the same file as procedure 11. If they are not in the same file, we'll consider completely different functions.
However, for aliens, this is not true. It is possible to have an inlined function that is defined from different file.
For example if function "bar" is inlined in file A, B and C, then we'll have 3 function "bar" in the procedure table.

For inline functions, hpcstruct may give a hint to hpcprof where the inline is defined (if the information is provided by the compiler), and it's up to hpcprof to interpret the information.
For instance:

<p i="1" n="foo">
...
</p>
<p ...>
 ...
  <a i="200" n="foo" ln="1">

This means that the node 200 which is inlined foo, is defined in procedure with ID=1. This is a hint for hpcprof that there's no need to create a separate flat ID for alien 200.

At the moment, hpcprof doesn't use this information yet. I am not sure how reliable this approach is. 
In some cases, some inline functions have no procedure definitions. Need further investigation.