Add two attributes for MetricDBTable xml tag: t and partner
These attributes are useful for hpcviewer to display metric values of
certain threads. If the root's metric is empty (no value), we can now
look at the value of the partner. Usually, the exclusive value of a root
is empty while the inclusive value is not.

Now, the xml of MetricDBTable is as follows:

  <MetricDBTable>
    <MetricDB i="0" n="REALTIME (usec) (I)" t="inclusive" partner="1"
db-glob="*.metric-db" db-id="0" db-num-metrics="2" db-header-sz="24"/>
    <MetricDB i="1" n="REALTIME (usec) (E)" t="exclusive" partner="0"
db-glob="*.metric-db" db-id="1" db-num-metrics="2" db-header-sz="24"/>
  </MetricDBTable>