Merge the hpctoolkit-parseapi branch back into trunk.
Both the old (OpenAnalysis) and new (ParseAPI) versions of hpcstruct
should be buildable with the current externals.  For now, it is a
compile-time option and the default is old (OA).  Use the config
option '--enable-parseapi' to build the new version.

Git complained that the hpctoolkit-parseapi branch was not a mergeable
object.  Maybe because the split was from svn.  So, I used the
merge(1) utility plus some scripts.  The result is the same, but git
will see this commit as having only one parent.