hpcprof: Minor change on how we compute exclusive cost for procedure frame.
Since now we consider inline functions (Proc class) the same as a normal procedure (ProcFrm class), we attribute the exclusive costs of statement of an inline function (Proc class) to the inline function itself instead of the enclosing procedure (ProcFrm class).
Previously, we attribute the exclusive cost of a statement to the enclosing procedure (the one which call the inline method) instead of the inline method itself.
For instance:

inline foo()
{
  for_loop();
}

bar() {
  foo();
}

Now, the exclusive cost of for_loop is attributed to foo() instead of bar().

Note: this update only has an impact in hpcprof when we have multiple hpcrun files, i.e. when hpcprof needs to compute the "derived incremental metrics". If there's only one hpcrun file in the measurement directory, it depends on hpcviewer to interpret the data.