doc www: add info to how to build hpcviewer. Now hpcviewer requires hpcviewer-externals plugin.
It is more appropriate to separate the source codes of external libraries into different repositories. This should ease merge with the original repository. Moreover, we don't want to claim the codes to be part of hpcviewer (or hpctoolkit).
Building hpcviewer with plugin from different repositories is a bit tricky. Right now, the easy way is to include the "external" repositories into the same Eclipse workspace as the viewer.