hpcprof: computing Mean with the aggregate NumSrc
Originally, the formula of Mean is Sum / NumSrc.
However, this results inconsistency in Callers view and Flat view
since NumSrc metric is accumulated for every time the scope is
merged.

Ideally, NumSrc shouldn't change between the 3 views.
We may need to have another combine function to avoid the
accumulation for NumSrc.