hpcprof:avoid accumulating NumSrc in combine fct
The goal is not to accumulate NumSrc metric while creating
Callers view and Flat view. Using sum() for the combine function
causes the accumulation. Hence, we just use '$x' for the combine
function so that the value of NumSrc (for Mean, StdDev and CoVar)
is still the same across the three views.