Update build tools to autoconf 2.69, automake 1.15 and libtool 2.4.6. This much just passes autoreconf, it won't actually build anything on the new platforms yet.
Add AC_LANG_SOURCE to configure.ac for IFELSE tests.  Autoconf prefers
this and generates a spew of warnings without it.

Add subdir-objects to a few Makefile.am.  Again, automake prefers this
and generates a spew of warnings without it.

Move libhpcrun.o from pkglib_PROGRAMS to pkglibexec_PROGRAMS.
In automake 1.15, pkglibdir is not a legal directory for programs.
This currently breaks the static case and we need to rework this.