fix hpcprof-mpi merging, which manifested itself as a "mergeDeep" failure.
when presented with imprecise data from binary analysis, where some
instruction addresses are attributable to source lines and others only
to procedures, hpcprof-mpi failed to merge CCTs properly.
measurement data attributed to a procedure but not a specific line
is mapped to the first line of a procedure. measurement data for
the first line of a procedure was not fused properly in this case
by hpcprof-mpi.