Simplify memory-overrides to call external "plugin"
Instead of registering dynamically a plugin which causes the code to be highly complex, we will statically register a plugin.
The advantage of this approach is simplicity;
while the limitation is only one "plugin" can be registered with memory-overrides.

Hence, we cannot have both memleak and data centric events. we need to pick either one.