Rework how we extend StmtInfo to a range of multiple instructions. Change StmtMap to a derived class of std::map and add a few methods.
This mostly finishes this step, except for extending the insert()
method to handle the case of one new interval spanning several
existing intervals.

This step is just preliminary work.  The real work on reducing memory
usage comes next.  Even so, this much reduces the memory to analyze
the hpgmg-fv binary (27 meg) from 1.0 gig to 725 meg.