Step 1 of repeal and replace for writing an hpcstruct file.  Repeal and eliminate using the Prof::Struct scope classes.  Replace with writing the struct file directly from the TreeNode format.  This is the first step to processing and printing the struct file one procedure at a time.
This much only produces the struct header, there are no files or
procedures.  But it does so without converting the TreeNode to a
Struct::Proc scope node.

This step also marks the end of using Open Analysis for hpcstruct.
OA will probably still build, but it certainly won't run and produce
a struct file.