enable specifying precise ip manually with hpcrun flags
Ideally, we should be able to setup precise_ip field in perf events. However, it is too complex to identify the type of the current running CPU.
Since parsing /proc/cpuinfo is not an ideal way within hpcrun, we just let users to specify which level of precision they want. If the precise level is not supported by the architecture, we just "unable to use the PMU".