Add clean-up procedures in case of segv
Now segv_handler.c provides a method to register a callback functions to
be called when a segv occurs. This method is called just before the long
jump so that any components can use the callback for cleaning-up the
messes.

In case of unwinding (ui_tree.c), the clean-up callback consists of
switching the state to NEVER.
Before ui_recipe_map calls build_intervals (which is prone to segv in
xed) we store the current the context (its state is FORTCOMING).
Once a segv occurs, the "clean-up" function will switch the state into
NEVER, and thus avoid the deadlock.