Working version of math expression in hpcprof
Proof of concept of parsing and evaluating math expression formulated by
hpcrun, computed in hpcprof and viewed in hpcviewer.
The current state is good enough to be used for computing simple math
formulae (like computing "adjusted metric due to multiplexing or memory
leak). Currently it is not designed for more advanced expression (log,
exp, sqrt, ....).

TODO:
- We need to generalize variable substitution class (VarMap) not only to
  substitute with metrics, but to other class.
- The result of math expression is always "double". Although VarMap
  class can handle both real and int metric values, ExprEval is not
  generic enough to return for other types.
- Helper metrics shouldn't be visible in hpcviewer.
- Need to inform users that "adjusted" metric values are not exact
  values. (case for multiplexing)