Adding a metric to adjust the counter due to multiplexing
Instead of using a formula computed in hpcprof, we have two metrics for
every events (when the value multiplexing flag is 1):

- one for original raw counter
- one for the "adjusted" counter which equals the raw counter multipled
  by scale factor (enabled time divided by running time)

Right now the default is to produce one metric, which is the original
raw counter.
The problem with perf events is that we don't know if the kernel will do
the multiplexing or not. We only know this when the sampling occurs and
we get the values of enable and running time.