Add frequency as a sampling "threshold"
Perf events allows to specify a frequency as a sampling threshold. This
means the period can vary dynamically as the kernel tries its best to
match with the specified frequency.

For a sample with frequency event, we need to increase the metric
counter with the period information from the kernel (a special flag is
needed to get the period information).

For period sampling, the counter is still increased by one.