Remove option to disable multiplexing
In perf events it is hard to tell kernel not to multiplex (the multiplex
is the default). To simplify the code, it's better to let the kernel
does the job.

Minor fix: allow all signal to be handled even if the si_code is not
POLL_HUP. The rational is some events (mostly software events) delivers
si_code with other flags than POLL_HUP.