Add metric info for: multiplexing and period
Multiplexed event can be inaccurate, especially if it's combined with
frequency-based sampling. It is important to notify users that the
metric is based on multiplexed event or not. If it's multiplexed (or
it's frequency-based sampling), we should inform the average of sampling
period and its standard deviation.

Minor update: reorganize the code for customized event (such as kernel
blocking)