Add PERF_RECORD_SWITCH to get info on context switches
A draft to include PERF_RECORD_SWITCH to get more info on context switches. This feature is only available on Linux Kernel 4.3.
For older kernels, we avoid using this feature.

At this commit, when profiling sleep unit test, we only get "context-out", without "context-in".
Need to investigate further why context-in disappears. Perhaps due to interference with sleep libc.