Stop setting, and then clearing, all the left children in a tree of intervals for x86.
bitree_uwi_finalize exists to clear the left children of nodes
along the right spine of a binary tree.  It is applied for x86
because we build an alleged binary tree that is really a doubly
linked list, and before we 'rebalance' it, we have to make it
an actual tree.

The new_ui function that make new tree nodes is passed an argument
that is a pointer to the previously allocated node.  On x86, this
is made the left child of the new node, until is it 'finalized'
back to NULL.  On ppc, this argument is ignored, and no 'finalization'
is needed.  Drop this argument from all calls to new_ui.