restructuring for "custom" events
Custom event such as kernel blocking or hierarchical events require clean infrastructure.
We need to separate between events defined in perfmon, custom event, and predefined hierarchical events (similar to perf tool's predefined events).

Also we may need to provide option for using perf event user callchain or not. At the moment the stack provided by kernel is not as "clean" as our built-in call stack. Need to be investigated.