Delete the old fashion of gathering context switch info
In the old way, we assume the first context switch event is the "enter" event, and the second context-switch event is the "leave" event.
However, this is not the case. On Linux 4.3 (and newer), we should use the attribute "context_switch" which then give us information which event is "enter" or "leave".

Although this approach works well if the monitoring process is not in the same process as the application, this doeesn't work for the first-party monitoring approach (hpctoolkit-way).
With the first-party monitoring, we only receives "context-out", and no "context-in".

Still needs to investigate further.