Using pre-allocated buffer to store call chains
Since we outsource the parsing of perf buffer to another file (and another function), we need a mechanism to temporarily allocate a variable lenght of array but cannot use alloca() anymore.
hpcrun_malloc is a good candidate, but it isn't suitable for temporary buffer since there's no free function.

A simplest solution is to use pre-allocated buffer, and make sure we don't store data more than the size of the buffer.