Restructure the code and selectively record callpath sample
Set guards for:
- when time cs out is bigger than current time
- when sample is really in the kernel by using PERF_RECORD_MISC_KERNEL
and kernel callpath. In the future we can use just
PERF_RECORD_MISC_KERNEL.

Minor update:
- ensure to sample if the header type is PERF_RECORD_SAMPLE. We don't
want to record the sample if the header type is THROTTLE/UNTHROTTLE or
others. In some tests, some times we encounter throttling, and since the
mmap buffer of this header is basically empty, I don't think it's
reliable to record the sample for this type (such as we'll have
incorrect period sampling). 
We need to decide what we should do in case of throttling. I saw only
some 700s throttles out of tens of thousands samples. Does it worth to
consider ?