Add generic support to display list of events
Instead of each forcing sample-sources to format their display of "list
of available events", it's better to provide generic API to format the
display.

The new support will automatically re-arrange a long description event
into multiple lines and align the column. It also handles a long event
name (which is common in native events).