Add additional struct to allow to extend hpcrun_sample_callpath
Adding information to hpcrun_sample_callpath is tricky since it's used by many parts of the code.
To reduce many changes due to adding another parameter to the function, it's more prudent to have an abstraction in the argument so that other part of the codes can add info without making a lot of changes.

minor update: guard in case uw_recipe_map_inrange_find() returns null (happens when profiling hpctruct (shell version) in perf branch.