Fix issues with nwchem and perf events
Somehow, nwchem can expose fragility of perf events:
- it triggers many POLL_IN even when the counter is not zero (overflows)
- using 4000 samples per second, it causes Linux kernel to set the
threshold to 22 cycles ! This causes infinite loop since once we finish
handling the event and enable the event, it immediately signals a new
event.

For the first case, we need to filter POLL_IN, and to turn off
"wakeup_events" attribute. According to the manpage, this attribute will
notify the handler before an overflow notification happens.
Since we use IOC_REFRESH, I don't see the need of this attribute.

For the second case, there's no fix yet.
However, if we disable TMSG, it works just fine, and there's no infinite
loop. 
On the other hand, if TMSG is enabled, somehow Linux kernel sets the
"adaptive" period threshold into 22 (which means every 22 cycles, the
handler is notified for a new event !) only if the hpcrun's number of
samples is 4000 per second. 
If the number of samples in hpcrun is 500 per second, the "adaptive"
period is correctly set (not as grazy as 22).